requirements.txt
streamlit==1.40.1
#opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.1.3
torch==2.5.1
torchvision==0.20.1
diffusers==0.31.0
transformers==4.46.3
scipy==1.14.1
#matplotlib==3.8.0
platformdirs==4.3.6
datasets==3.1.0
ipywidgets==8.1.5